pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "codearkt"
version = "1.9.5"
description = "CodeAct implementation with Docker and MCP"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "Ilya Gusev", email = "phoenixilya@gmail.com"}
]
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]
dependencies = [
    "academia-mcp>=1.11.1",
    "arize-phoenix-otel>=0.12.1",
    "docker>=7.1.0",
    "fastmcp==2.11.3",
    "httpx>=0.27.0",
    "jinja2>=3.1.6",
    "openai>=1.93.0",
    "openinference-instrumentation-openai>=0.1.30",
    "pytest-asyncio>=1.0.0",
    "types-docker>=7.1.0.20250523",
    "types-pyyaml>=6.0.12.20250516",
    "wsproto>=1.2.0",
    "uvicorn>=0.34.3",
    "sse-starlette<3.0.0",
    "mcp>=1.10.1",
    "gradio==5.29.0",
    "tiktoken>=0.11.0",
    "prompt-toolkit>=3.0.52",
    "a2a-sdk[http-server]>=0.3.8",
]

[dependency-groups]
dev = [
    "black==25.1.0",
    "mypy==1.16.0",
    "flake8==7.2.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=7.0.0",
    "isort>=6.1.0",
]

[project.urls]
Homepage = "https://github.com/IlyaGusev/codearkt"

[tool.setuptools.packages.find]
where = ["."]
include = ["codearkt*"]
exclude = ["container*"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
strict = true
explicit_package_bases = true

[[tool.mypy.overrides]]
module = "pydantic.*"
follow_imports = "skip"

[[tool.mypy.overrides]]
module = "fastapi.*"
ignore_missing_imports = true
follow_imports = "skip"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_test_loop_scope = "function"

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
"codearkt" = ["prompts/*.yaml"]
